Refactor: all-reduce the CSA output projection with an atomic-add window
Each rank now dequantizes its own partial locally and atomic-adds the FP32
result into one shared band, instead of publishing a raw INT32 partial plus
its activation scale into a per-rank slot and summing the eight slots after
the barrier.

- oproj_dequant folds both scales before the wire. The per-channel wo_b_scale
  distributes over the cross-rank sum, so it rides that pass too and the band
  holds the finished value.
- The publish drops the separate [T, 1] scale put: 16 transfers become 8.
- oproj_reduce narrows the band to BF16 and zeroes the lane it read.
- oproj_band_init zeroes both lanes on a dispatch's first call and barriers,
  because alloc_window_buffer does not promise zeroed memory and the reduce's
  zero-back has no predecessor there. Trip count is 1 only at FIRST_EPOCH.
- The reduce window is now [2 * T_PAD, D] FP32 rather than
  [2 * N_RANKS * T_PAD, D] INT32; decode_csa.py and decode_layer.py allocate
  the same window and change with it.
- DEQUANT_N_TILE and REDUCE_D_TILE are 1024 (4 blocks each): both grids are
  element-light and per-block fixed-cost bound, so they want fewer, wider
  blocks than the eight-slot reduce they replace.

Behaviour change: the cross-rank sum order is now whatever the atomics land
in, so the result is no longer bit-reproducible run to run, nor term-for-term
equal to a single card's g loop. attn_out, x_next and kv_cache PASS on the
standalone and decode_layer goldens.

This costs latency and is taken deliberately. a2a3, 8 cards, one frozen
golden, fastest rank: standalone (--pre-sync, 100 rounds / 5 warmup, median)
62.6 -> 65.0 us; decode_layer --layer-id 2 (50 rounds / 5 warmup) median
632.4 -> 640.2 us, min 613.5 -> 620.2. Deleting the scale put is worth ~7.6 us
of publish span and the reduce drops from 183.6 to 35.2 us of core time, but
the added dequant pass and the first-call band init more than spend both.

+++ models/deepseek_v4_flash_lowlat/decode_csa.py
@@ -377,7 +377,7 @@ def attention_csa(
     wo_b_shard: pl.Tensor[[D, O_LORA], pl.INT8],
     wo_b_scale: pl.Tensor[[D], pl.FP32],
     x_out: pl.Tensor[[T, HC_MULT, D], pl.FP32],
-    reduce_window: pld.DistributedTensor[[OPROJ_REDUCE_ROWS, D], pl.INT32],
+    reduce_window: pld.DistributedTensor[[OPROJ_REDUCE_ROWS, D], pl.FP32],
     scale_window: pld.DistributedTensor[[OPROJ_SCALE_ROWS, 1], pl.FP32],
     reduce_signal: pld.DistributedTensor[[OPROJ_N_RANKS, 1], pl.INT32],
     sync_signal: pld.DistributedTensor[[OPROJ_N_RANKS, 1], pl.INT32],
@@ -477,7 +477,7 @@ def attention_csa_test(
     wo_b_shard: pl.Tensor[[D, O_LORA], pl.INT8],
     wo_b_scale: pl.Tensor[[D], pl.FP32],
     x_out: pl.Out[pl.Tensor[[T, HC_MULT, D], pl.FP32]],
-    reduce_window: pld.DistributedTensor[[OPROJ_REDUCE_ROWS, D], pl.INT32],
+    reduce_window: pld.DistributedTensor[[OPROJ_REDUCE_ROWS, D], pl.FP32],
     scale_window: pld.DistributedTensor[[OPROJ_SCALE_ROWS, 1], pl.FP32],
     reduce_signal: pld.DistributedTensor[[OPROJ_N_RANKS, 1], pl.INT32],
     sync_signal: pld.DistributedTensor[[OPROJ_N_RANKS, 1], pl.INT32],
@@ -546,13 +546,13 @@ def l3_attention_csa(
     x_out: pl.Out[pl.Tensor[[OPROJ_N_RANKS, T, HC_MULT, D], pl.FP32]],
 ):
     """One orchestration per card, sharing the reduce and signal windows."""
-    reduce_window_buf = pld.alloc_window_buffer([OPROJ_REDUCE_ROWS, D], dtype=pl.INT32)
+    reduce_window_buf = pld.alloc_window_buffer([OPROJ_REDUCE_ROWS, D], dtype=pl.FP32)
     scale_window_buf = pld.alloc_window_buffer([OPROJ_SCALE_ROWS, 1], dtype=pl.FP32)
     reduce_signal_buf = pld.alloc_window_buffer([OPROJ_N_RANKS, 1], dtype=pl.INT32)
     sync_signal_buf = pld.alloc_window_buffer([OPROJ_N_RANKS, 1], dtype=pl.INT32)
 
     for r in pl.range(pld.world_size()):
-        reduce_window = pld.window(reduce_window_buf, [OPROJ_REDUCE_ROWS, D], dtype=pl.INT32)
+        reduce_window = pld.window(reduce_window_buf, [OPROJ_REDUCE_ROWS, D], dtype=pl.FP32)
         scale_window = pld.window(scale_window_buf, [OPROJ_SCALE_ROWS, 1], dtype=pl.FP32)
         reduce_signal = pld.window(reduce_signal_buf, [OPROJ_N_RANKS, 1], dtype=pl.INT32)
         sync_signal = pld.window(sync_signal_buf, [OPROJ_N_RANKS, 1], dtype=pl.INT32)
